Written a lot, also added report.pdf (for fun)
To remove report.pdf just use git rm --cached report.pdf
In other news: Drafted conclusion, discussion and security chapter.
Lots of other small changes here and there, mostly additions and formatting.

Also had a lot of fun setting up a git hook on the mac so that everytime I pushed to it
a local version of the repo was pulled, a pdf was built and opened :D
This commit represents 48 instances of git commit -am test && git push
<3 Git